% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{ggmiNEXT3D}
\alias{ggmiNEXT3D}
\title{ggplot2 extension for an miNEXT3D object}
\usage{
ggmiNEXT3D(out)
}
\arguments{
\item{out}{an \code{miNEXT3D} object computed by \code{\link{miNEXT3D}}.}
}
\value{
a ggplot2 object
}
\description{
\code{ggmiNEXT3D}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{miNEXT3D}} Object to plot sized-based rarefaction/extrapolation curves
}
\examples{
# diversity = 'TD'
data(dunes)
data <- dunes$data
data1 <- data[,c(2,1)]
output1 <- miNEXT3D(data1, diversity = 'TD', nboot = 10)
plot1 <- ggmiNEXT3D(output1)
plot1

# diversity = 'PD'
data(dunes)
data <- dunes$data
data1 <- data[,c(2,1)]
tree <- dunes$tree
output2 <- miNEXT3D(data1, diversity = 'PD', PDtree = tree, nboot = 10)
plot2 <- ggmiNEXT3D(output2)
plot2

# diversity = 'FD' & FDtype = 'tau_values'
data(dunes)
data <- dunes$data
data1 <- data[,c(2,1)]
distM <- as.matrix(dunes$dist)
output3 <- miNEXT3D(data1, diversity = 'FD', FDdistM = distM, FDtau = 0.3, FDtype = 'tau_values', nboot = 10)
plot3 <- ggmiNEXT3D(output3)
plot3

# diversity = 'FD' & FDtype = 'AUC'
data(dunes)
data <- dunes$data
data1 <- data[,c(2,1)]
distM <- as.matrix(dunes$dist)
output4 <- miNEXT3D(data1, diversity = 'FD', FDdistM = distM, FDtype = 'AUC', nboot = 0)
plot4 <- ggmiNEXT3D(output4)
plot4

}
